Extract srlookup function into own file, move JSDOM scraping into it to make it easier to unit-test
